requirements.txt
flask==1.1.2
requests==2.24.0
google-cloud-storage==1.31.2
google-cloud-bigquery==1.28.0
google-cloud-language==1.3.0
pytube3==9.6.4
werkzeug==1.0.1
